test: server tests
